rewrite DescriptionCollectionAssertion to ...Expectation
