+++ translations/de_CH/atrium-translations-de_CH-common/src/main/kotlin/ch/tutteli/atrium/translations/deprecatedDescriptionCharSequenceAssertion.kt
@@ -1,3 +1,6 @@
+//TODO remove file with 1.0.0 at the latest
+@file:Suppress("DEPRECATION")
+
 package ch.tutteli.atrium.translations
 
 import ch.tutteli.atrium.assertions.DescriptiveAssertion
@@ -6,6 +9,7 @@ import ch.tutteli.atrium.reporting.translating.StringBasedTranslatable
 /**
  * Contains the [DescriptiveAssertion.description]s of the assertion functions which are applicable to [CharSequence].
  */
+@Deprecated("Will be removed with 1.0.0 at the latest")
 enum class DescriptionCharSequenceAssertion(override val value: String) : StringBasedTranslatable {
     @Deprecated(
         "Use DescriptionCharSequenceExpectation instead; will be removed with 1.0.0 at the latest",
