+++ translations/en_GB/atrium-translations-en_GB-common/src/main/kotlin/ch/tutteli/atrium/translations/DescriptionCollectionExpectation.kt
@@ -1,0 +1,15 @@
+package ch.tutteli.atrium.translations
+
+import ch.tutteli.atrium.assertions.DescriptiveAssertion
+import ch.tutteli.atrium.reporting.translating.StringBasedTranslatable
+
+/**
+ * Contains the [DescriptiveAssertion.description]s of the expectation functions which are applicable to [Collection].
+ */
+enum class DescriptionCollectionExpectation(override val value: String) : StringBasedTranslatable {
+    /** @since 0.18.0 */
+    EMPTY("empty"),
+
+    /** @since 0.18.0 */
+    SIZE("size"),
+}
